refactor(runtime): consolidate ready queue protocol
